reorganize, appease clippy
